Database sync now resyncs every now and then (30 secs in debug and 2 hours in prod).
Fixed the bug of the same monitor message re-sent again and again. In case the same row was updated, the rowid changes, causing the row to not be found (by rowid). In case this happens, we will just ignore the message and move on, since the subsequent hook trigger will take care of this update.