What the fuck.
So many changes. How am I supposed to describe this in the commit message.

I guess this is a partial rewrite of the registry. So many things cleaned up. So much more cleanup to do. I fucking hate this. My only solace is that once this is implemented I won't have to touch this again. Fuck this